+++ src/extension/Services/CodeActionProviderExtractLabel.ts
@@ -1,12 +1,12 @@
-import { CodeAction, CodeActionKind, Location, Range, TextDocument, TextEdit, TextLine, WorkspaceEdit } from "vscode";
+import { CodeAction, CodeActionKind, commands, Location, Range, SnippetString, TextDocument, TextEdit, TextLine, window, workspace, WorkspaceEdit } from "vscode";
 import { FullSyntaxTreeNodeKind } from '../AL Code Outline Ext/fullSyntaxTreeNodeKind';
-import { TextRangeExt } from '../AL Code Outline Ext/textRangeExt';
+import { TextRangeExt } from "../AL Code Outline Ext/textRangeExt";
 import { ALFullSyntaxTreeNode } from '../AL Code Outline/alFullSyntaxTreeNode';
 import { SyntaxTree } from '../AL Code Outline/syntaxTree';
 import { ALVariable } from '../Entities/alVariable';
 import { Command } from '../Entities/Command';
 import { Config } from '../Utils/config';
-import { DocumentUtils } from '../Utils/documentUtils';
+import { DocumentUtils } from "../Utils/documentUtils";
 import { LabelComment } from "../Utils/labelComment";
 import { WorkspaceEditUtils } from '../Utils/WorkspaceEditUtils';
 import { ICodeActionProvider } from "./ICodeActionProvider";
@@ -32,28 +32,39 @@ export class CodeActionProviderExtractLabel implements ICodeActionProvider {
             return [];
 
         let stringLiteralRange: Range = DocumentUtils.trimRange(this.document, TextRangeExt.createVSCodeRange(this.stringLiteralTreeNode.fullSpan));
-        let extractLabelCreatesComment: boolean = Config.getExtractToLabelCreatesComment(this.document.uri);
-        let commentText: string;
-        if (extractLabelCreatesComment) {
-           commentText = LabelComment.getCommentTextForLabel(this.document, stringLiteralRange);
-        } else {
-           commentText = ''
-        }
-        let variable: ALVariable = new ALVariable('newLabel', 'Label ' + this.document.getText(stringLiteralRange) + commentText);
-        let textEdit: TextEdit = WorkspaceEditUtils.addVariableToLocalVarSection(methodOrTriggerTreeNode, variable, this.document);
-
-        let edit: WorkspaceEdit = new WorkspaceEdit();
-        edit.replace(this.document.uri, stringLiteralRange, variable.name!);
-        edit.insert(this.document.uri, textEdit.range.start, textEdit.newText);
-        let linesAdded: number = (textEdit.newText.length - textEdit.newText.replace(/\r\n/g, '').length) / 2;
-
         let codeAction: CodeAction = new CodeAction('Extract to Label', CodeActionKind.RefactorExtract);
-        codeAction.edit = edit;
         codeAction.command = {
-            command: Command.renameCommand,
-            arguments: [new Location(this.document.uri, stringLiteralRange.start.translate(linesAdded))],
+            command: Command.extractLabel,
+            arguments: [this.document, this.range, stringLiteralRange, methodOrTriggerTreeNode],
             title: 'Extract Label'
         };
         return [codeAction];
+    }
+    public async runCommand(stringLiteralRange: Range, methodOrTriggerTreeNode: ALFullSyntaxTreeNode) {
+        let extractLabelCreatesComment: boolean = Config.getExtractToLabelCreatesComment(this.document.uri);
+        let commentText: string = '';
+        if (extractLabelCreatesComment)
+            commentText = LabelComment.getCommentTextForLabel(this.document, stringLiteralRange);
+        let snippetMode: boolean = commentText != '';
+
+        let cleanVariableName = 'newLabel';
+        let edit: WorkspaceEdit = new WorkspaceEdit();
+        edit.replace(this.document.uri, stringLiteralRange, cleanVariableName);
+
+        let variableName = cleanVariableName;
+        if (snippetMode)
+            variableName = '${0:' + cleanVariableName + '}';
+        let variable: ALVariable = new ALVariable(variableName, 'Label ' + this.document.getText(stringLiteralRange) + commentText);
+        let textEdit: TextEdit = WorkspaceEditUtils.addVariableToLocalVarSection(methodOrTriggerTreeNode, variable, this.document);
+
+        if (snippetMode) {
+            await workspace.applyEdit(edit);
+            await window.activeTextEditor!.insertSnippet(new SnippetString(textEdit.newText), textEdit.range.start, { undoStopBefore: false, undoStopAfter: false })
+        } else {
+            edit.insert(this.document.uri, textEdit.range.start, textEdit.newText);
+            await workspace.applyEdit(edit);
+            let linesAdded: number = (textEdit.newText.length - textEdit.newText.replace(/\r\n/g, '').length) / 2;
+            await commands.executeCommand(Command.renameCommand, new Location(this.document.uri, stringLiteralRange.start.translate(linesAdded)));
+        }
     }
 }