+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to the "al-codeactions" extension will be documented in this
 
 - fix #128: Extract procedure should allow rename even if a comment is inside the selection
 - fix #126: Position of new procedures: If the "anchor" after which the new procedure should be added is a trigger, then the new procedure should be added after the last global variable section - not the first one
+- Add configuration "Extract to Label Creates Comment": This will add a Comment section with the used placeholders to explain them.
 
 ## 1.0.7